+++ src/ecc/codes/qeccs_using_oscar.jl
@@ -1,6 +1,9 @@
 """Implemented in a package extension with `Oscar`. Check the docs for the [Oscar extension](http://qc.quantumsavory.org/stable/ECC_API/#Implemented-in-an-extension-requiring-Oscar.jl)"""
 function boundary_maps end
 
+"""Implemented in a package extension with `Oscar`. Check the docs for the [Oscar extension](http://qc.quantumsavory.org/stable/ECC_API/#Implemented-in-an-extension-requiring-Oscar.jl)"""
+function max_xy_exponents end
+
 """D-Dimensional Surface codes ([Berthusen_2024](@cite), [Zeng_2019](@cite))
 
 Implemented as a package extension with `Oscar`. Check the [QuantumClifford documentation](http://qc.quantumsavory.org/stable/ECC_API/) for more details on that extension."""
@@ -43,4 +46,15 @@ function DoubleHomologicalProductCode(args...; kwargs...)
         throw("The `DoubleHomologicalProductCode` depends on the package `Oscar` but you have not installed or imported it yet. Immediately after you import `Oscar`, the `DoubleHomologicalProductCode` will be available.")
     end
     return ext.DoubleHomologicalProductCode(args...; kwargs...)
+end
+
+"""Generalized Toric codes ([liang2025generalizedtoriccodestwisted](@cite))
+
+Implemented as a package extension with `Oscar`. Check the [QuantumClifford documentation](http://qc.quantumsavory.org/stable/ECC_API/) for more details on that extension."""
+function GeneralizedToricCode(args...; kwargs...)
+    ext = Base.get_extension(QuantumClifford, :QuantumCliffordOscarExt)
+    if isnothing(ext)
+        throw("The `GeneralizedToricCode` depends on the package `Oscar` but you have not installed or imported it yet. Immediately after you import `Oscar`, the `GeneralizedToricCode` will be available.")
+    end
+    return ext.GeneralizedToricCode(args...; kwargs...)
 end